+++ src/CDA.Generator.Common/SCSModel/Common/Entities/Frequency.cs
@@ -84,6 +84,8 @@ namespace Nehta.VendorLibrary.CDA.SCSModel.Common
               narrative += !Denominator.Units.IsNullOrEmptyWhitespace() ? string.Format(" {0}", Denominator.Units)  : String.Empty;              
             }
 
+
+
             if (Numerator != null)
             {
               narrative += !Numerator.Value.IsNullOrEmptyWhitespace() ? string.Format(" {0}", Numerator.Value) : String.Empty;