+++ src/CDA.R5Sample/PathologyReportWithStructuredContentSample.cs
@@ -508,11 +508,13 @@ namespace CDA.R5Samples
             resultGroup.ResultValue = BaseCDAModel.CreateResultValue();
 
             // Test Result Value
-            resultGroup.ResultValue.TestResultValue = BaseCDAModel.CreateQuantity("12.88", "ml");
-            //resultGroup.ResultValue.ValueAsQuantityRangeValueAsQuantityRange = BaseCDAModel.CreateQuantityRange(1.2, 1.1, "ml");
+            //resultGroup.ResultValue.TestResultValue = BaseCDAModel.CreateQuantity("12.88", "ml");
+            //resultGroup.ResultValue.ValueAsQuantityRange = BaseCDAModel.CreateQuantityRange(1.2, 1.1, "ml");
+            resultGroup.ResultValue.ValueAsQuantityRange = BaseCDAModel.CreateQuantityRange(1.1, 1.1, "ml");
             //resultGroup.ResultValue.ValueAsCodableText = BaseCDAModel.CreateCodableText("75540009", CodingSystem.SNOMEDCT, "High");
             //resultGroup.ResultValue.ValueAsBoolean = true;
             //resultGroup.ResultValue.ValueAsString = "Result in upper range";
+            //resultGroup.ResultValue.ValueAsInteger = 1;
             //resultGroup.ResultValue.ValueAsRatio = BaseCDAModel.CreateRatio("10", "ml", "25", "ml");
 
             if (!mandatorySectionsOnly)